Clock out of org before ending test. Note follows:
The test yogurty-test/clocked-into-rt-ticket-sad-path-not-clocked-in was
failing, but only if run in after
yogurty-test/clocked-into-rt-ticket-number-only-new-ticket.  If run on
its own, it worked; if run as part of big batch it failed; if run after
clocked-into, it failed.

Adding this checkout to the fixture appears to fix this problem; now
20/20 tests pass, rather than 19/20.  I can also run just these two
tests and they pass.

However, I'm still suspicious.  It occurs to me that is a case where
the *test code itself* is complicated enough, or mysterious enough (for
me, anyhow, a newbie at Lisp and this test harness) that I'm starting to
wonder if the test code needs a test suite itself.  This is either the
beginning of wisdom or the limit of sanity.